TG: Delete - Edit Icon, Warning message for delete
