+++ src/Components/Employees/List/employees-list.module.css
@@ -1,0 +1,13 @@
+table {
+    margin: 40px 25px;
+    min-width: 100%;
+    table-layout: auto;
+    border-collapse: collapse;
+  }
+  tr, th {
+    border: 1px solid rgb(128, 128, 128);
+  }
+  th {
+    padding: 10px 5px;
+    text-align: start;
+  }
