+++ src/Components/Employees/ListItem/employees-item.module.css
@@ -1,0 +1,17 @@
+.button {
+    align-self: flex-end;
+    padding: 5px 8px;
+    margin: 5px 4px;
+    border: 1px solid rgb(24, 24, 24);
+    border-radius: 5px;
+    color: rgb(24, 24, 24);
+    background-color: rgb(240, 244, 249);
+    font-size: 16px;
+    text-decoration: none;
+}
+
+.button:hover {
+    background-color: rgb(255, 255, 255);
+    transform: scale(1.05);
+    transition: all 0.2s ease-in-out;
+}
